Initial Commit - Create Address Service
Address Service provides a http service for fetching addresses
using a hint. It is built in Rust, and utilized an inverted
index for storing addresses.

An address file, addresses.json, provides all of the known addresses.
There are no mechanisms for updating addresses, because if there are
new addresses, they are simply added to addresses.json and the
service is redeployed.

The service runs locally on port 8000, and is accessed like so:

`http://localhost:8000/addresses?hint=<the_address_hint>`

Additionally, there is an optional limit parameter, `limit`,
that provides the ability to specify the number of results to
return. The default is 3.